Update naming to match main target
